Make bucketing algorithm faster (20 sec to .12 sec on sample data)
Before it was O(k * n) where k was the # of buckets and n was the number of vehicles.

Most speed studies are done over a number of days, so you end up with a lot of
buckets with reasonable bucket intervals (5 to 60 minutes).

The new algorithm is O(n) with the bucket names generated on the fly.